requirements.txt
aiosqlite==0.17.0
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
black==21.9b0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.0
flake8==4.0.1
h11==0.12.0
httptools==0.2.0
idna==3.3
iso8601==0.1.16
isort==5.9.3
mccabe==0.6.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.4.0
PyJWT==2.3.0
pypika-tortoise==0.1.1
python-decouple==3.5
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
regex==2021.10.23
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tomli==1.2.1
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
